+++ react-tutorial/src/Square.js
@@ -1,0 +1,8 @@
+function Square({ value, onSquareClick }) {
+  return (
+    <button className="square" onClick={onSquareClick}>
+      {value}
+    </button>
+  );
+}
+export default Square;
